+++ docs/changes.md
@@ -1,6 +1,9 @@
 Changes
 =======
 
+#### 0.0.4
+* replace `JSONEncoderExt` with `JSONEncoderPlus` (flask-based)
+
 #### 0.0.3
 * remove module level `import flask` in `j.f.serialize`
 * add Decimal support in `j.f.ser.JSONEncoderExt`